Remove ADDRESS_SS58 from Precompile extension and clean up
