+++ precompiles/src/balance_transfer.rs
@@ -1,17 +1,17 @@
 use core::marker::PhantomData;
 
 use frame_support::dispatch::{GetDispatchInfo, PostDispatchInfo};
+use frame_system::RawOrigin;
 use pallet_evm::PrecompileHandle;
 use precompile_utils::EvmResult;
 use sp_core::{H256, U256};
 use sp_runtime::traits::{Dispatchable, StaticLookup, UniqueSaturatedInto};
 
-use crate::parser::{contract_to_origin, parse_pubkey};
 use crate::{PrecompileExt, PrecompileHandleExt};
 
 pub(crate) struct BalanceTransferPrecompile<R>(PhantomData<R>);
 
-impl<R> PrecompileExt for BalanceTransferPrecompile<R>
+impl<R> PrecompileExt<R::AccountId> for BalanceTransferPrecompile<R>
 where
     R: frame_system::Config + pallet_balances::Config + pallet_evm::Config,
     R::AccountId: From<[u8; 32]>,
@@ -24,11 +24,6 @@ where
     <R as pallet_balances::Config>::Balance: TryFrom<U256>,
 {
     const INDEX: u64 = 2048;
-    const ADDRESS_SS58: Option<[u8; 32]> = Some([
-        0x07, 0xec, 0x71, 0x2a, 0x5d, 0x38, 0x43, 0x4d, 0xdd, 0x03, 0x3f, 0x8f, 0x02, 0x4e, 0xcd,
-        0xfc, 0x4b, 0xb5, 0x95, 0x1c, 0x13, 0xc3, 0x08, 0x5c, 0x39, 0x9c, 0x8a, 0x5f, 0x62, 0x93,
-        0x70, 0x5d,
-    ]);
 }
 
 #[precompile_utils::precompile]
@@ -53,18 +48,13 @@ where
             return Ok(());
         }
 
-        let dest = parse_pubkey::<R::AccountId>(address.as_bytes())?.0.into();
+        let dest = R::AccountId::from(address.0).into();
 
         let call = pallet_balances::Call::<R>::transfer_allow_death {
             dest,
             value: amount_sub.unique_saturated_into(),
         };
 
-        handle.try_dispatch_runtime_call::<R, _>(
-            call,
-            contract_to_origin(
-                &Self::ADDRESS_SS58.expect("ADDRESS_SS58 is defined for BalanceTransferPrecompile"),
-            )?,
-        )
+        handle.try_dispatch_runtime_call::<R, _>(call, RawOrigin::Signed(Self::address_ss58()))
     }
 }
