+++ precompiles/src/ed25519.rs
@@ -1,21 +1,26 @@
 extern crate alloc;
 
 use alloc::vec::Vec;
+use core::marker::PhantomData;
 
 use ed25519_dalek::{Signature, Verifier, VerifyingKey};
 use fp_evm::{ExitError, ExitSucceed, LinearCostPrecompile, PrecompileFailure};
 
 use crate::PrecompileExt;
-use crate::parser::parse_slice;
 
-pub(crate) struct Ed25519Verify;
+pub(crate) struct Ed25519Verify<A>(PhantomData<A>);
 
-impl PrecompileExt for Ed25519Verify {
+impl<A> PrecompileExt<A> for Ed25519Verify<A>
+where
+    A: From<[u8; 32]>,
+{
     const INDEX: u64 = 1026;
-    const ADDRESS_SS58: Option<[u8; 32]> = None;
 }
 
-impl LinearCostPrecompile for Ed25519Verify {
+impl<A> LinearCostPrecompile for Ed25519Verify<A>
+where
+    A: From<[u8; 32]>,
+{
     const BASE: u64 = 15;
     const WORD: u64 = 3;
 
@@ -45,5 +50,23 @@ impl LinearCostPrecompile for Ed25519Verify {
         };
 
         Ok((ExitSucceed::Returned, buf.to_vec()))
+    }
+}
+
+/// Takes a slice from bytes with PrecompileFailure as Error
+fn parse_slice(data: &[u8], from: usize, to: usize) -> Result<&[u8], PrecompileFailure> {
+    let maybe_slice = data.get(from..to);
+    if let Some(slice) = maybe_slice {
+        Ok(slice)
+    } else {
+        log::error!(
+            "fail to get slice from data, {:?}, from {}, to {}",
+            &data,
+            from,
+            to
+        );
+        Err(PrecompileFailure::Error {
+            exit_status: ExitError::InvalidRange,
+        })
     }
 }
