+++ precompiles/src/staking.rs
@@ -37,7 +37,6 @@ use sp_core::{H256, U256};
 use sp_runtime::traits::{Dispatchable, StaticLookup, UniqueSaturatedInto};
 use subtensor_runtime_common::ProxyType;
 
-use crate::parser::{parse_pubkey, try_u16_from_u256};
 use crate::{PrecompileExt, PrecompileHandleExt};
 
 // Old StakingPrecompile had ETH-precision in values, which was not alligned with Substrate API. So
@@ -47,7 +46,7 @@ use crate::{PrecompileExt, PrecompileHandleExt};
 // All the future extensions should happen in StakingPrecompileV2.
 pub(crate) struct StakingPrecompileV2<R>(PhantomData<R>);
 
-impl<R> PrecompileExt for StakingPrecompileV2<R>
+impl<R> PrecompileExt<R::AccountId> for StakingPrecompileV2<R>
 where
     R: frame_system::Config
         + pallet_evm::Config
@@ -62,7 +61,6 @@ where
     <<R as frame_system::Config>::Lookup as StaticLookup>::Source: From<R::AccountId>,
 {
     const INDEX: u64 = 2053;
-    const ADDRESS_SS58: Option<[u8; 32]> = None;
 }
 
 #[precompile_utils::precompile]
@@ -90,7 +88,7 @@ where
     ) -> EvmResult<()> {
         let account_id = handle.caller_account_id::<R>();
         let amount_staked = amount_rao.unique_saturated_into();
-        let (hotkey, _) = parse_pubkey(address.as_bytes())?;
+        let hotkey = R::AccountId::from(address.0);
         let netuid = try_u16_from_u256(netuid)?;
         let call = pallet_subtensor::Call::<R>::add_stake {
             hotkey,
@@ -109,7 +107,7 @@ where
         netuid: U256,
     ) -> EvmResult<()> {
         let account_id = handle.caller_account_id::<R>();
-        let (hotkey, _) = parse_pubkey(address.as_bytes())?;
+        let hotkey = R::AccountId::from(address.0);
         let netuid = try_u16_from_u256(netuid)?;
         let amount_unstaked = amount_alpha.unique_saturated_into();
         let call = pallet_subtensor::Call::<R>::remove_stake {
@@ -126,7 +124,7 @@ where
         _handle: &mut impl PrecompileHandle,
         coldkey: H256,
     ) -> EvmResult<U256> {
-        let (coldkey, _) = parse_pubkey(coldkey.as_bytes())?;
+        let coldkey = R::AccountId::from(coldkey.0);
         let stake = pallet_subtensor::Pallet::<R>::get_total_stake_for_coldkey(&coldkey);
 
         Ok(stake.into())
@@ -135,9 +133,9 @@ where
     #[precompile::public("getTotalHotkeyStake(bytes32)")]
     fn get_total_hotkey_stake(
         _handle: &mut impl PrecompileHandle,
-        hotkey_h256: H256,
+        hotkey: H256,
     ) -> EvmResult<U256> {
-        let (hotkey, _) = parse_pubkey(hotkey_h256.as_bytes())?;
+        let hotkey = R::AccountId::from(hotkey.0);
         let stake = pallet_subtensor::Pallet::<R>::get_total_stake_for_hotkey(&hotkey);
 
         Ok(stake.into())
@@ -151,8 +149,8 @@ where
         coldkey: H256,
         netuid: U256,
     ) -> EvmResult<U256> {
-        let (hotkey, _) = parse_pubkey(hotkey.as_bytes())?;
-        let (coldkey, _) = parse_pubkey(coldkey.as_bytes())?;
+        let hotkey = R::AccountId::from(hotkey.0);
+        let coldkey = R::AccountId::from(coldkey.0);
         let netuid = try_u16_from_u256(netuid)?;
         let stake = pallet_subtensor::Pallet::<R>::get_stake_for_hotkey_and_coldkey_on_subnet(
             &hotkey, &coldkey, netuid,
@@ -164,7 +162,7 @@ where
     #[precompile::public("addProxy(bytes32)")]
     fn add_proxy(handle: &mut impl PrecompileHandle, delegate: H256) -> EvmResult<()> {
         let account_id = handle.caller_account_id::<R>();
-        let (delegate, _) = parse_pubkey(delegate.as_bytes())?;
+        let delegate = R::AccountId::from(delegate.0);
         let delegate = <R as frame_system::Config>::Lookup::unlookup(delegate);
         let call = pallet_proxy::Call::<R>::add_proxy {
             delegate,
@@ -178,7 +176,7 @@ where
     #[precompile::public("removeProxy(bytes32)")]
     fn remove_proxy(handle: &mut impl PrecompileHandle, delegate: H256) -> EvmResult<()> {
         let account_id = handle.caller_account_id::<R>();
-        let (delegate, _) = parse_pubkey(delegate.as_bytes())?;
+        let delegate = R::AccountId::from(delegate.0);
         let delegate = <R as frame_system::Config>::Lookup::unlookup(delegate);
         let call = pallet_proxy::Call::<R>::remove_proxy {
             delegate,
@@ -193,7 +191,7 @@ where
 // Deprecated, exists for backward compatibility.
 pub(crate) struct StakingPrecompile<R>(PhantomData<R>);
 
-impl<R> PrecompileExt for StakingPrecompile<R>
+impl<R> PrecompileExt<R::AccountId> for StakingPrecompile<R>
 where
     R: frame_system::Config
         + pallet_evm::Config
@@ -211,11 +209,6 @@ where
     <<R as frame_system::Config>::Lookup as StaticLookup>::Source: From<R::AccountId>,
 {
     const INDEX: u64 = 2049;
-    const ADDRESS_SS58: Option<[u8; 32]> = Some([
-        0x26, 0xf4, 0x10, 0x1e, 0x52, 0xb7, 0x57, 0x34, 0x33, 0x24, 0x5b, 0xc3, 0x0a, 0xe1, 0x8b,
-        0x63, 0x99, 0x53, 0xd8, 0x41, 0x79, 0x33, 0x03, 0x61, 0x4d, 0xfa, 0xcf, 0xf0, 0x37, 0xf7,
-        0x12, 0x94,
-    ]);
 }
 
 #[precompile_utils::precompile]
@@ -247,7 +240,7 @@ where
         }
 
         let amount_sub = handle.try_convert_apparent_value::<R>()?;
-        let (hotkey, _) = parse_pubkey(address.as_bytes())?;
+        let hotkey = R::AccountId::from(address.0);
         let netuid = try_u16_from_u256(netuid)?;
         let call = pallet_subtensor::Call::<R>::add_stake {
             hotkey,
@@ -266,7 +259,7 @@ where
         netuid: U256,
     ) -> EvmResult<()> {
         let account_id = handle.caller_account_id::<R>();
-        let (hotkey, _) = parse_pubkey(address.as_bytes())?;
+        let hotkey = R::AccountId::from(address.0);
         let netuid = try_u16_from_u256(netuid)?;
         let amount_unstaked =
             <R as pallet_evm::Config>::BalanceConverter::into_substrate_balance(amount)
@@ -284,9 +277,9 @@ where
     #[precompile::public("getTotalColdkeyStake(bytes32)")]
     fn get_total_coldkey_stake(
         _handle: &mut impl PrecompileHandle,
-        coldkey_h256: H256,
+        coldkey: H256,
     ) -> EvmResult<U256> {
-        let (coldkey, _) = parse_pubkey(coldkey_h256.as_bytes())?;
+        let coldkey = R::AccountId::from(coldkey.0);
 
         // get total stake of coldkey
         let total_stake = pallet_subtensor::Pallet::<R>::get_total_stake_for_coldkey(&coldkey);
@@ -301,9 +294,9 @@ where
     #[precompile::public("getTotalHotkeyStake(bytes32)")]
     fn get_total_hotkey_stake(
         _handle: &mut impl PrecompileHandle,
-        hotkey_h256: H256,
+        hotkey: H256,
     ) -> EvmResult<U256> {
-        let (hotkey, _) = parse_pubkey(hotkey_h256.as_bytes())?;
+        let hotkey = R::AccountId::from(hotkey.0);
 
         // get total stake of hotkey
         let total_stake = pallet_subtensor::Pallet::<R>::get_total_stake_for_hotkey(&hotkey);
@@ -323,8 +316,8 @@ where
         coldkey: H256,
         netuid: U256,
     ) -> EvmResult<U256> {
-        let (hotkey, _) = parse_pubkey(hotkey.as_bytes())?;
-        let (coldkey, _) = parse_pubkey(coldkey.as_bytes())?;
+        let hotkey = R::AccountId::from(hotkey.0);
+        let coldkey = R::AccountId::from(coldkey.0);
         let netuid = try_u16_from_u256(netuid)?;
         let stake = pallet_subtensor::Pallet::<R>::get_stake_for_hotkey_and_coldkey_on_subnet(
             &hotkey, &coldkey, netuid,
@@ -339,7 +332,7 @@ where
     #[precompile::public("addProxy(bytes32)")]
     fn add_proxy(handle: &mut impl PrecompileHandle, delegate: H256) -> EvmResult<()> {
         let account_id = handle.caller_account_id::<R>();
-        let (delegate, _) = parse_pubkey(delegate.as_bytes())?;
+        let delegate = R::AccountId::from(delegate.0);
         let delegate = <R as frame_system::Config>::Lookup::unlookup(delegate);
         let call = pallet_proxy::Call::<R>::add_proxy {
             delegate,
@@ -353,7 +346,7 @@ where
     #[precompile::public("removeProxy(bytes32)")]
     fn remove_proxy(handle: &mut impl PrecompileHandle, delegate: H256) -> EvmResult<()> {
         let account_id = handle.caller_account_id::<R>();
-        let (delegate, _) = parse_pubkey(delegate.as_bytes())?;
+        let delegate = R::AccountId::from(delegate.0);
         let delegate = <R as frame_system::Config>::Lookup::unlookup(delegate);
         let call = pallet_proxy::Call::<R>::remove_proxy {
             delegate,
@@ -368,9 +361,6 @@ where
         account_id: &<R as frame_system::Config>::AccountId,
         amount: U256,
     ) -> Result<(), PrecompileFailure> {
-        let smart_contract_account_id = R::AccountId::from(
-            Self::ADDRESS_SS58.expect("ADDRESS_SS58 is defined for StakingPrecompile"),
-        );
         let amount_sub =
             <R as pallet_evm::Config>::BalanceConverter::into_substrate_balance(amount)
                 .ok_or(ExitError::OutOfFund)?;
@@ -385,7 +375,7 @@ where
 
         // Execute the transfer
         let transfer_result =
-            transfer_call.dispatch(RawOrigin::Signed(smart_contract_account_id).into());
+            transfer_call.dispatch(RawOrigin::Signed(Self::address_ss58()).into());
 
         if let Err(dispatch_error) = transfer_result {
             log::error!(
@@ -399,4 +389,10 @@ where
 
         Ok(())
     }
+}
+
+fn try_u16_from_u256(value: U256) -> Result<u16, PrecompileFailure> {
+    value.try_into().map_err(|_| PrecompileFailure::Error {
+        exit_status: ExitError::Other("the value is outside of u16 bounds".into()),
+    })
 }
